pyproject.toml
[tool.poetry]
name = "cif3r"
version = "0.3.0"
description = "A classifier to determine whether an object is recycleable"
authors = ["deterministic-forest <jwilli18@go.olemiss.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
Pillow = "^7.0.0"
selenium = "^3.141.0"
pandas = "^0.25.3"
requests = "^2.22.0"
matplotlib = "^3.1.2"
sklearn = "^0.0"
scikit-learn = "^0.22.1"
setuptools = "^41.0.0"
flask = "^1.1.1"
flask_sqlalchemy = "^2.4.1"
opencv-python = "^4.1.2"
seaborn = "^0.10.0"
networkx = "^2.4"
tqdm = "^4.42.1"
torch = "^1.5.0"
torchvision = "^0.6.0"
streamlit = "^0.63.1"
torchsummary = "^1.5.1"
protobuf = "^3.12.2"
tensorboard = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
black = "^19.10b0"
pylint = "^2.4.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
